// packages/alloy-typespec/openapi3-to-typespec/src/openapi-types.ts
export type Extensions = {
  [key in `x-${string}`]?: any;
};

export interface Ref {
  $ref: string;
  description?: string;
  summary?: string;
}

export type Refable<T> = Ref | T;
// ...
export type OpenAPISchema3_1 = JsonSchema<
  {
    /**
     * Adds support for polymorphism.
     * The discriminator is an object name that is used to differentiate between other schemas which may satisfy the payload description
     *
     * @see https://spec.openapis.org/oas/v3.1.1.html#discriminator-object
     */
    discriminator?: OpenAPI3Discriminator;

    /**
     * Additional external documentation for this schema.
     *
     * @see https://spec.openapis.org/oas/v3.1.1.html#external-documentation-object
     */
    externalDocs?: OpenAPI3ExternalDocs;

    /**
     * This MAY be used only on properties schemas.
     * It has no effect on root schemas.
     * Adds additional metadata to describe the XML representation of this property.
     */
    xml?: OpenAPI3XmlSchema;
  } & Extensions
>;

export interface OpenAPI3Discriminator extends Extensions {
  propertyName: string;
  mapping?: Record<string, string>;
}

export interface OpenAPI3ExternalDocs {
  url: string;
  description?: string;
}

export type OpenAPI3XmlSchema = Extensions & {
  name?: string;
  namespace?: string;
  prefix?: string;
  attribute?: boolean;
  wrapped?: boolean;
};

export type OpenAPI3Link =
  | {
      /** A relative or absolute reference to an OAS operation. This field is mutually exclusive of the operationId field, and MUST point to an Operation Object. Relative operationRef values MAY be used to locate an existing Operation Object in the OpenAPI definition.  */
      operationRef: Ref;
    }
  | {
      /** the name of an existing, resolvable OAS operation, as defined with a unique operationId. This field is mutually exclusive of the operationRef field. */
      operationId: string;
    };

export type OpenAPI3StatusCode =
  | string
  | "default"
  | "1XX"
  | "2XX"
  | "3XX"
  | "4XX"
  | "5XX";

export type OpenAPI3Responses = {
  [status: OpenAPI3StatusCode]: Refable<OpenAPI3Response>;
} & Extensions;

export type OpenAPI3Response = Extensions & {
  /** A short description of the response. CommonMark syntax MAY be used for rich text representation. */
  description: string;

  /** Maps a header name to its definition. RFC7230 states header names are case insensitive. If a response header is defined with the name "Content-Type", it SHALL be ignored. */
  headers?: Record<string, Refable<OpenAPI3Header>>;

  /** A map containing descriptions of potential response payloads. The key is a media type or media type range and the value describes it. For responses that match multiple keys, only the most specific key is applicable. e.g. text/plain overrides text/* */
  content?: Record<string, OpenAPI3MediaType>;

  /** A map of operations links that can be followed from the response. The key of the map is a short name for the link, following the naming constraints of the names for Component Objects. */
  links?: Record<string, Refable<OpenAPI3Link>>;
};
// ...
export type OpenAPI3SecurityScheme =
  | OpenAPI3ApiKeySecurityScheme
  | OpenAPI3OAuth2SecurityScheme
  | OpenAPI3OpenIdConnectSecurityScheme
  | OpenAPI3HttpSecurityScheme;

export interface OpenAPIComponents3_1 extends Extensions {
  schemas?: Record<string, Refable<OpenAPISchema3_1>>;
  responses?: Record<string, Refable<OpenAPI3Response>>;
  parameters?: Record<string, Refable<OpenAPI3Parameter>>;
  examples?: Record<string, Refable<OpenAPI3Example>>;
  requestBodies?: Record<string, Refable<OpenAPI3RequestBody>>;
  headers?: Record<string, Refable<OpenAPI3Header>>;
  securitySchemes?: Record<string, Refable<OpenAPI3SecurityScheme>>;
  links?: Record<string, Refable<OpenAPI3Link>>;
  callbacks?: Record<string, Record<string, OpenAPI3PathItem>>;
  pathItems?: Record<string, OpenAPI3PathItem>;
}

export interface OpenAPI3Info extends Extensions {
  title: string;
  description?: string;
  termsOfService?: string;
  version: string;
  contact?: unknown;
  license?: unknown;
  summary?: string;
}

export interface OpenAPI3Server {
  url: string;
  description?: string;
  variables?: Record<string, OpenAPI3ServerVariable>;
}

export interface OpenAPI3ServerVariable {
  enum?: string[];
  default: string;
  description?: string;
}

export interface OpenAPI3Tag extends Extensions {
  name: string;
  description?: string;
  externalDocs?: OpenAPI3ExternalDocs;
}

export type HttpMethod =
  | "get"
  | "put"
  | "post"
  | "delete"
  | "options"
  | "head"
  | "patch"
  | "trace";

/**
 * Describes the operations available on a single path. A Path Item may be empty, due to ACL constraints. The path itself is still exposed to the documentation viewer but they will not know which operations and parameters are available.
 *
 * @see https://github.com/OAI/OpenAPI-Specification/blob/master/versions/2.0.md#pathItemObject
 */
export type OpenAPI3PathItem = {
  [method in HttpMethod]?: OpenAPI3Operation;
} & { parameters?: Refable<OpenAPI3Parameter>[] } & Extensions;
// ...
export type OpenAPI3Parameter =
  | OpenAPI3HeaderParameter
  | OpenAPI3QueryParameter
  | OpenAPI3PathParameter;
// ...
export type OpenAPI3Header = OpenAPI3ParameterBase & {
  /** Describes how the parameter value will be serialized depending on the type of the parameter value
   *
   * simple - Simple style parameters defined by RFC6570. This option replaces collectionFormat with a csv value from OpenAPI 2.0
   */
  style?: "simple";
};
// ...
export interface OpenAPI3Example {
  summary?: string;
  description?: string;
  value?: unknown;
  externalValue?: string;
}

/**
 * Describes a single request body.
 *
 * @see https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.2.md#request-body-object
 */
export type OpenAPI3RequestBody = Extensions & {
  /** A brief description of the request body. This could contain examples of use. CommonMark syntax MAY be used for rich text representation. */
  description?: string;

  /** The content of the request body. The key is a media type or media type range and the value describes it. For requests that match multiple keys, only the most specific key is applicable. e.g. text/plain overrides text/* */
  content: Record<string, OpenAPI3MediaType>;

  /** Determines if the request body is required in the request. Defaults to false. */
  required?: boolean;
};

export type OpenAPI3Operation = Extensions & {
  description?: string;
  summary?: string;
  responses?: any;
  tags?: string[];
  operationId?: string;
  requestBody?: OpenAPI3RequestBody;
  parameters: Refable<OpenAPI3Parameter>[];
  deprecated?: boolean;
  security?: Record<string, string[]>[];
};

export interface OpenAPIDocument3_1 extends Extensions {
  openapi: "3.1.0";

  /**
   * Provides metadata about the API. The metadata can be used by the clients if needed.
   */
  info: OpenAPI3Info;

  /**
   * The default value for the $schema keyword within Schema Objects contained within this document.
   * This MUST be in the form of a URI.
   */
  jsonSchemaDialect?: string;

  /** Additional external documentation. */
  externalDocs?: OpenAPI3ExternalDocs;

  /** The available paths and operations for the API */
  paths: Record<string, Refable<OpenAPI3PathItem>>;

  /**
   * The incoming webhooks that may be received as part of this API.
   * The key name is a unique string to refer to each webhook.
   */
  webhooks?: Record<string, OpenAPI3PathItem>;

  /**
   * An array of Server Objects, which provide connectivity information to a target server.
   * If the servers property is not provided, or is an empty array, the default value would be a Server Object with a url value of /.
   */
  servers?: OpenAPI3Server[];

  /**
   * A list of tags used by the specification with additional metadata.
   * The order of the tags can be used to reflect on their order by the parsing tools.
   * Not all tags that are used by the Operation Object must be declared.
   * The tags that are not declared MAY be organized randomly or based on the tools' logic. Each tag name in the list MUST be unique.
   */
  tags?: OpenAPI3Tag[];

  /** An element to hold various schemas for the specification. */
  components?: OpenAPIComponents3_1;

  /**
   * A declaration of which security mechanisms can be used across the API.
   * The list of values includes alternative security requirement objects that can be used.
   * Only one of the security requirement objects need to be satisfied to authorize a request.
   * Individual operations can override this definition.
   */
  security?: Record<string, string[]>[];
}